fix B3 tests
