+++ propagator/b3/test/single_test.rb
@@ -7,20 +7,11 @@
 require 'test_helper'
 
 describe OpenTelemetry::Propagator::B3::Single do
-  describe '#text_map_extractor' do
-    it 'returns an instance of TextMapExtractor' do
-      extractor = OpenTelemetry::Propagator::B3::Single.text_map_extractor
-      _(extractor).must_be_instance_of(
-        OpenTelemetry::Propagator::B3::Single::TextMapExtractor
-      )
-    end
-  end
-
-  describe '#text_map_injector, #rack_injector' do
-    it 'returns an instance of TextMapInjector' do
-      injector = OpenTelemetry::Propagator::B3::Single.text_map_injector
-      _(injector).must_be_instance_of(
-        OpenTelemetry::Propagator::B3::Single::TextMapInjector
+  describe '#text_map_propagator' do
+    it 'returns an instance of TextMapPropagator' do
+      propagator = OpenTelemetry::Propagator::B3::Single.text_map_propagator
+      _(propagator).must_be_instance_of(
+        OpenTelemetry::Propagator::B3::Single::TextMapPropagator
       )
     end
   end
